requirements.txt
edgartools>=5.30.0
openai>=1.40.0
chromadb>=0.5.0
fastapi>=0.111.0
uvicorn>=0.30.0
jinja2>=3.1.0
mcp>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=8.0.0
tiktoken>=0.7.0
httpx>=0.27.0
anyio>=4.0.0
